a bit more debugging support
It appears there are race conditions that are caught by the
most recently added test case which are irreconcileable
under the current update polling. The issue arises due to
SSDs completing file operations too rapidly and a lack of
adequate timing resolution for file updates in the OS. The
final test case occasionally catches this issue by writing
a large block of data to a file and then rewriting with a
very small block. In such cases, Bopdrox will sometimes
catch the large write in progress, and then subsequently
miss the small update if it completes within the same
timing window.

This will unfortunately remain a problem until we finally
hook into the OS to receive updates.